Take float values

+++ script.js
@@ -16,8 +16,8 @@ function divide(a,b) {
 
 // calls an operator function based on input
 function operate(operator, num1, num2) {
-  num1 = parseInt(num1);
-  num2 = parseInt(num2);
+  num1 = parseFloat(num1);
+  num2 = parseFloat(num2);
   let result;
   switch (operator) {
     case "+":
